+++ spring-boot-project/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/mongo/ReactiveMongoClientFactory.java
@@ -33,9 +33,25 @@ import org.springframework.core.env.Environment;
  */
 public class ReactiveMongoClientFactory extends MongoClientFactorySupport<MongoClient> {
 
+	/**
+	 * Construct a factory for creating a {@link MongoClient}.
+	 * @param properties configuration properties
+	 * @param environment a Spring {@link Environment} containing configuration properties
+	 * @param builderCustomizers a list of configuration settings customizers
+	 * @deprecated since 2.4.0 in favor of {@link #ReactiveMongoClientFactory(List)}
+	 */
+	@Deprecated
 	public ReactiveMongoClientFactory(MongoProperties properties, Environment environment,
 			List<MongoClientSettingsBuilderCustomizer> builderCustomizers) {
 		super(properties, environment, builderCustomizers, MongoClients::create);
+	}
+
+	/**
+	 * Construct a factory for creating a {@link MongoClient}.
+	 * @param builderCustomizers a list of configuration settings customizers
+	 */
+	public ReactiveMongoClientFactory(List<MongoClientSettingsBuilderCustomizer> builderCustomizers) {
+		super(builderCustomizers, MongoClients::create);
 	}
 
 }
