+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/mongo/MongoClientFactoryTests.java
@@ -21,7 +21,6 @@ import java.util.List;
 import com.mongodb.MongoClientSettings;
 import com.mongodb.client.MongoClient;
 
-import org.springframework.core.env.Environment;
 import org.springframework.test.util.ReflectionTestUtils;
 
 /**
@@ -36,9 +35,9 @@ import org.springframework.test.util.ReflectionTestUtils;
 class MongoClientFactoryTests extends MongoClientFactorySupportTests<MongoClient> {
 
 	@Override
-	protected MongoClient createMongoClient(MongoProperties properties, Environment environment,
-			List<MongoClientSettingsBuilderCustomizer> customizers, MongoClientSettings settings) {
-		return new MongoClientFactory(properties, environment, customizers).createMongoClient(settings);
+	protected MongoClient createMongoClient(List<MongoClientSettingsBuilderCustomizer> customizers,
+			MongoClientSettings settings) {
+		return new MongoClientFactory(customizers).createMongoClient(settings);
 	}
 
 	@Override
